Unify file watch data sources
