+++ src/Microsoft.VisualStudio.ProjectSystem.Managed/ProjectSystem/AdditionalDesignTimeBuildInputDataSource.cs
@@ -1,71 +1,26 @@
-﻿ // Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
+﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
-using System;
-using System.Collections.Immutable;
 using System.ComponentModel.Composition;
-using System.Linq;
-using System.Threading.Tasks.Dataflow;
-using Microsoft.VisualStudio.ProjectSystem.Properties;
 
 namespace Microsoft.VisualStudio.ProjectSystem
 {
     /// <summary>
     ///     Responsible for producing AdditionalDesignTimeBuildInput items (such as project.assets.json)
-    ///     so that they can be watched, and when changed trigger re-evaluation.
+    ///     so that they can be watched, and when changed, trigger re-evaluation.
     /// </summary>
     [Export(typeof(IFileWatchDataSource))]
     [AppliesTo(ProjectCapability.PackageReferences)]
-    internal class AdditionalDesignTimeBuildInputDataSource : ChainedProjectValueDataSourceBase<FileWatchData>, IFileWatchDataSource
+    internal class AdditionalDesignTimeBuildInputDataSource : AbstractItemFileWatchDataSource
     {
-        private readonly ConfiguredProject _project;
-        private readonly IProjectSubscriptionService _projectSubscriptionService;
-
         [ImportingConstructor]
         public AdditionalDesignTimeBuildInputDataSource(ConfiguredProject project, IProjectSubscriptionService projectSubscriptionService)
-            : base(project.Services, synchronousDisposal:true, registerDataSource: false)
+            : base(project, projectSubscriptionService)
         {
-            _project = project;
-            _projectSubscriptionService = projectSubscriptionService;
         }
 
-        protected override UnconfiguredProject ContainingProject
-        {
-            get { return _project.UnconfiguredProject; }
-        }
+        public override string ItemRuleName => AdditionalDesignTimeBuildInput.SchemaName;
 
-        protected override IDisposable LinkExternalInput(ITargetBlock<IProjectVersionedValue<FileWatchData>> targetBlock)
-        {
-            IProjectValueDataSource<IProjectSubscriptionUpdate> source = _projectSubscriptionService.ProjectRuleSource;
+        public override FileWatchChangeKinds FileWatchChangeKinds => FileWatchChangeKinds.Added | FileWatchChangeKinds.Removed | FileWatchChangeKinds.Changed;
 
-            // Transform the changes from evaluation/design-time build -> restore data
-            DisposableValue<ISourceBlock<IProjectVersionedValue<FileWatchData>>> transformBlock = source.SourceBlock
-                                                                                                        .TransformWithNoDelta(update => update.Derive(u => CreateFileWatch(u.CurrentState)),
-                                                                                                            suppressVersionOnlyUpdates: true,
-                                                                                                            ruleNames: new[] { AdditionalDesignTimeBuildInput.SchemaName });
-
-            // Set the link up so that we publish changes to target block
-            transformBlock.Value.LinkTo(targetBlock, DataflowOption.PropagateCompletion);
-
-            // Join the source blocks, so if they need to switch to UI thread to complete 
-            // and someone is blocked on us on the same thread, the call proceeds
-            JoinUpstreamDataSources(source);
-
-            return transformBlock;
-        }
-
-        private FileWatchData CreateFileWatch(IImmutableDictionary<string, IProjectRuleSnapshot> update)
-        {
-            IProjectRuleSnapshot snapshot = update.GetSnapshotOrEmpty(AdditionalDesignTimeBuildInput.SchemaName);
-
-            var fullPaths = snapshot.Items.Select(item => item.Value)
-                                          .Select(metadata => metadata.GetValueOrDefault(AdditionalDesignTimeBuildInput.FullPathProperty))
-                                          .Where(item => item != null)
-                                          .ToImmutableList();
-
-            return new FileWatchData(
-                this,
-                fullPaths,
-                FileWatchChangeKinds.Added | FileWatchChangeKinds.Removed | FileWatchChangeKinds.Changed);
-        }
     }
 }
